fix:protocal handshake probelms fix
- Update the mgrCtx checks to allow it to be NULL.
- Fix the encryptThenMac handling during session resumption.
- Fix the wraparound issue when receiving client extensions.

Cherry-picked from: https://gitcode.com/openHiTLS/openhitls/merge_requests/1446

+++ include/tls/hitls.h
@@ -1686,7 +1686,7 @@ int32_t HITLS_ClientHelloGetCiphers(HITLS_Ctx *ctx, uint16_t **out, uint16_t *ou
 * @retval  HITLS_SUCCESS, if successful.
 *          For details about other error codes, see hitls_error.h.
     */
-int32_t HITLS_ClientHelloGetExtensionsPresent(HITLS_Ctx *ctx, uint16_t **out, uint8_t *outlen);
+int32_t HITLS_ClientHelloGetExtensionsPresent(HITLS_Ctx *ctx, uint16_t **out, uint32_t *outlen);
 
 /**
  * @ingroup tls
